+++ lib/lightning_web/live/credential_live/json_schema_body_component.ex
@@ -8,7 +8,9 @@ defmodule LightningWeb.CredentialLive.JsonSchemaBodyComponent do
 
   def fieldset(%{form: form} = assigns) do
     changeset = form.source
-    schema = changeset |> Ecto.Changeset.get_field(:schema) |> get_schema()
+
+    schema =
+      changeset |> Ecto.Changeset.get_field(:schema) |> Credentials.get_schema()
 
     schema_changeset =
       create_schema_changeset(
@@ -63,21 +65,6 @@ defmodule LightningWeb.CredentialLive.JsonSchemaBodyComponent do
       </div>
     </fieldset>
     """
-  end
-
-  # false positive, it's a file from config
-  # sobelow_skip ["Traversal.FileModule"]
-  defp get_schema(schema_name) do
-    {:ok, schemas_path} = Application.fetch_env(:lightning, :schemas_path)
-
-    File.read("#{schemas_path}/#{schema_name}.json")
-    |> case do
-      {:ok, raw_json} ->
-        Credentials.Schema.new(raw_json, schema_name)
-
-      {:error, reason} ->
-        raise "Error reading credential schema. Got: #{reason |> inspect()}"
-    end
   end
 
   defp create_schema_changeset(schema, params) do